feat: implement view offset editing with double-precision support and UI integration

+++ src/main/java/shipeditor/components/viewer/layers/ship/data/ShipHull.java
@@ -40,7 +40,7 @@ public class ShipHull {
 
     private Color coversColor;
 
-    private int viewOffset;
+    private double viewOffset;
 
     private String hullFileName;
